[Delivers #160752374] rename models folder to app
