+++ run.py
@@ -1,4 +1,4 @@
-from models.api import app
+from app.api import app
 
 if __name__ == '__main__':
     app.run(debug=True)
